Implement items tests (#6)

+++ src/main/scala/com/kevchuang/shop/sql/codecs.scala
@@ -2,14 +2,14 @@ package com.kevchuang.shop.sql
 
 import com.kevchuang.shop.domain.brand.*
 import com.kevchuang.shop.domain.category.*
+import com.kevchuang.shop.domain.currency.*
 import com.kevchuang.shop.domain.item.*
 import com.kevchuang.shop.domain.price.*
-import com.kevchuang.shop.domain.currency.*
 import com.kevchuang.shop.domain.types.common.NotEmpty
-import skunk.*
-import skunk.codec.all.*
 import io.github.iltotore.iron.*
 import io.github.iltotore.iron.constraint.all.*
+import skunk.*
+import skunk.codec.all.*
 
 import java.util.UUID
 
@@ -31,8 +31,8 @@ object codecs:
     varchar.eimap(_.refineEither[NotEmpty].map(ItemName(_)))(_.value)
 
   val price: Codec[Price] =
-    float8
-      .eimap(a => a.refineEither[Positive].map(e => USD(Amount(e))))(
+    numeric
+      .eimap(a => a.toDouble.refineEither[Positive].map(e => USD(Amount(e))))(
         _.amount.value
       )
